+++ src/main/java/jeff/storage/JeffException.java
@@ -1,3 +1,5 @@
+package jeff.storage;
+
 public class JeffException extends Exception {
     public JeffException(String message) {
         super(message);
