+++ src/main/java/jeff/task/Event.java
@@ -1,3 +1,5 @@
+package jeff.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
