+++ src/main/java/jeff/task/Task.java
@@ -1,3 +1,5 @@
+package jeff.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
