+++ src/main/java/jeff/ui/UserInterface.java
@@ -1,3 +1,5 @@
+package jeff.ui;
+
 import java.util.Scanner;
 
 public class UserInterface {
